+++ packages/cli/src/commands/dev/generator/index.ts
@@ -1,5 +1,6 @@
 import fs from "fs";
 import path from "path";
+import chalk from "chalk";
 
 import { FormGenerateDto } from "./dto";
 
@@ -14,8 +15,41 @@ export class Generator {
 
   form(dto: FormGenerateDto) {
     fs.writeFileSync(
-      path.join(this.generatedDirectory, `${dto.name.replace(" ", "-")}.tsx`),
-      dto.code,
+      path.join(this.generatedDirectory, `${dto.fileName}.tsx`),
+      dto.code
     );
+  }
+
+  delete(fileName: string) {
+    const filePath = path.join(this.generatedDirectory, `${fileName}.tsx`);
+
+    fs.unlink(filePath, (err) => {
+      if (err) {
+        console.error(`Error removing file: ${err.message}`);
+      } else {
+        chalk.gray(`File ${fileName} has been successfully removed.`);
+      }
+    });
+  }
+
+  rename(fileNameBefore: string, fileNameAfter: string) {
+    const sourceFilePath = path.join(
+      this.generatedDirectory,
+      `${fileNameBefore}.tsx`
+    );
+    const destinationFilePath = path.join(
+      this.generatedDirectory,
+      `${fileNameAfter}.tsx`
+    );
+
+    fs.rename(sourceFilePath, destinationFilePath, (err) => {
+      if (err) {
+        console.error(`Error moving file: ${err.message}`);
+      } else {
+        console.log(
+          chalk.gray(`-- Form renamed: ${fileNameBefore} -> ${fileNameAfter}`)
+        );
+      }
+    });
   }
 }