+++ packages/react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @lezzform/react
 
+## 0.1.0
+
+### Minor Changes
+
+- form support with action from platform
+
+### Patch Changes
+
+- Updated dependencies
+  - @lezzform/types@0.0.3
+
 ## 0.0.4
 
 ### Patch Changes