+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lezzform/types
 
+## 0.0.3
+
+### Patch Changes
+
+- form action type initial
+
 ## 0.0.2
 
 ### Patch Changes